requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclass-wizard==0.22.2
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
novita_client==0.4.10
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
torch==2.1.2
torchvision==0.16.2
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
